support and refactor into absolute inputs
