+++ src/App/Auth/Login/Login.scss
@@ -1,4 +1,5 @@
-@import '../../../shared.scss';
+@import '_transitions.scss';
+@import '_card.scss';
 
 .login-page {
   width: 100vw;
@@ -24,20 +25,20 @@
 
       .lock-container {
         position: absolute;
-        transition: 0.35s cubic-bezier(0.165, 0.84, 0.44, 1);
+        transition: $track-transition;
         transform: translateX(-95px);
       }
 
       .text-container {
         font-size: 20px;
         font-weight: 700;
-        transition: 0.35s cubic-bezier(0.165, 0.84, 0.44, 1);
+        transition: $track-transition;
         transform: translateX(10px);
       }
 
       .heart-container {
         position: absolute;
-        transition: 0.35s cubic-bezier(0.165, 0.84, 0.44, 1);
+        transition: $track-transition;
         transform: translateX(85px);
         opacity: 0;
       }
