+++ src/App/Auth/PrivateRoute/index.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import Login from '../Login'
 import Loader from '../../Loader'
 import getLoaderMessage from '../../Loader/sillyExcuses'
-import { ConsumerContainer } from '../../Contexts'
+import { ConsumerContainer } from 'Contexts/index'
 import { initSpotifyClient, requestNewToken } from '../spotify'
 
 const PrivateRouteContainer = props => (
