+++ src/App/Playlist/LoadedPlaylist/Main/Tracklist/TrackContainer/TrackContainer.scss
@@ -1,4 +1,4 @@
-@import '../../../../../../shared.scss';
+@import 'styles.scss';
 
 $track-transition: 0.35s cubic-bezier(0.165, 0.84, 0.44, 1);
