+++ src/App/Search/Search.scss
@@ -1,4 +1,4 @@
-@import '../../shared.scss';
+@import 'styles.scss';
 
 .search-page {
   width: 100vw;
