+++ src/_card.scss
@@ -1,5 +1,3 @@
-$track-transition: 0.35s cubic-bezier(0.165, 0.84, 0.44, 1);
-
 .card {
   background: white;
   border-radius: 4px;
