+++ src/_transitions.scss
@@ -1,0 +1,1 @@
+$track-transition: 0.35s cubic-bezier(0.165, 0.84, 0.44, 1);
