+++ src/styles.scss
@@ -1,0 +1,2 @@
+@import 'transitions';
+@import 'card';
